+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Craft Commerce
 
+## 4.2.8 - 2023-05-03
+
+- Added `craft\commerce\services\Customers::EVENT_UPDATE_PRIMARY_PAYMENT_SOURCE`.
+- Fixed a bug where PDFs could be generated using the wrong formatting locale. ([#3145](https://github.com/craftcms/commerce/issues/3145))
+
 ## 4.2.7 - 2023-04-13
 
 - Added the “Order Site” order condition rule. ([#3131](https://github.com/craftcms/commerce/issues/3131))